+++ CHANGELOG.md
@@ -22,6 +22,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - [DesignToolbox] Redirection page for stable builds in about page (Orange-OpenSource/ouds-ios#1457)
+- [Library] Rich text for `alert message`, `bullet list` components (Orange-OpenSource/ouds-ios#1405)
+
+### Fixed
+
 - [Library] Overlay items of `tab bar` component still displayed even if hidden for disabled Liquid Glass (Orange-OpenSource/ouds-ios#1434)
 - [Library] Selected tab indicator in `tab bar` component if Liquid Glass not applied (Orange-OpenSource/ouds-ios#1428)